requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
python-dotenv>=1.0.0
dj-database-url>=2.1
psycopg2-binary>=2.9
gunicorn>=21.0
whitenoise>=6.5
google-genai>=1.0.0
